add window, components
