+++ source/Apps/FrontendMentor.ProductPreviewCardComponent/App.xaml
@@ -1,3 +1,14 @@
-﻿<Application x:Class="FrontendMentor.ProductPreviewCardComponent.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" />
+﻿<applications:FrontendMentorApplication x:Class="FrontendMentor.ProductPreviewCardComponent.App"
+                                        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                                        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                                        xmlns:applications="clr-namespace:FrontendMentor.Assets.Applications;assembly=FrontendMentor.Assets"
+                                        x:ClassModifier="internal">
+    <applications:FrontendMentorApplication.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="pack://application:,,,/FrontendMentor.Assets;component/Themes/Custom.xaml" />
+                <ResourceDictionary Source="pack://application:,,,/FrontendMentor.ProductPreviewCardComponent;component/Themes/Custom.xaml" />
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </applications:FrontendMentorApplication.Resources>
+</applications:FrontendMentorApplication>
