+++ source/Apps/FrontendMentor.ProductPreviewCardComponent/Constants/ProductPreviewCardComponentViewNames.cs
@@ -1,0 +1,17 @@
+﻿// --------------------------------------------------------------------------------
+// Copyright (C) 2025 Eugene Eremeev (also known as Yevhenii Yeriemeieiv).
+// All Rights Reserved.
+// --------------------------------------------------------------------------------
+// This software is the confidential and proprietary information of Eugene Eremeev
+// (also known as Yevhenii Yeriemeieiv) ("Confidential Information"). You shall not
+// disclose such Confidential Information and shall use it only in accordance with
+// the terms of the license agreement you entered into with Eugene Eremeev (also
+// known as Yevhenii Yeriemeieiv).
+// --------------------------------------------------------------------------------
+
+namespace FrontendMentor.ProductPreviewCardComponent.Constants;
+
+internal static class ProductPreviewCardComponentViewNames
+{
+    public const string ProductPreviewCardComponent = nameof(ProductPreviewCardComponent);
+}
