+++ source/Apps/FrontendMentor.ProductPreviewCardComponent/Themes/Generic.xaml
@@ -1,0 +1,5 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:resourceDictionaries="clr-namespace:FrontendMentor.Assets.ResourceDictionaries;assembly=FrontendMentor.Assets">
+    <ResourceDictionary.MergedDictionaries>
+        <resourceDictionaries:SharedResourceDictionary Source="pack://application:,,,/FrontendMentor.ProductPreviewCardComponent;component/Styles/Windows/ProductPreviewCardComponentWindowStyles.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+</ResourceDictionary>
