+++ source/Apps/FrontendMentor.ProductPreviewCardComponent/ViewModels/ProductPreviewCardComponentViewModel.cs
@@ -1,0 +1,18 @@
+﻿// --------------------------------------------------------------------------------
+// Copyright (C) 2025 Eugene Eremeev (also known as Yevhenii Yeriemeieiv).
+// All Rights Reserved.
+// --------------------------------------------------------------------------------
+// This software is the confidential and proprietary information of Eugene Eremeev
+// (also known as Yevhenii Yeriemeieiv) ("Confidential Information"). You shall not
+// disclose such Confidential Information and shall use it only in accordance with
+// the terms of the license agreement you entered into with Eugene Eremeev (also
+// known as Yevhenii Yeriemeieiv).
+// --------------------------------------------------------------------------------
+
+using FrontendMentor.Core.ViewModels;
+
+namespace FrontendMentor.ProductPreviewCardComponent.ViewModels;
+
+internal sealed class ProductPreviewCardComponentViewModel : NavigationViewModelBase
+{
+}
