+++ source/Apps/FrontendMentor.ProductPreviewCardComponent/Views/ProductPreviewCardComponentView.xaml
@@ -1,0 +1,9 @@
+﻿<UserControl x:Class="FrontendMentor.ProductPreviewCardComponent.Views.ProductPreviewCardComponentView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:viewModels="clr-namespace:FrontendMentor.ProductPreviewCardComponent.ViewModels"
+             d:DataContext="{d:DesignInstance viewModels:ProductPreviewCardComponentViewModel}"
+             x:ClassModifier="internal"
+             mc:Ignorable="d" />
