+++ source/Apps/FrontendMentor.QrCodeComponent/Styles/Windows/QrCodeComponentWindowStyles.xaml
@@ -1,9 +1,9 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:assetsWindows="clr-namespace:FrontendMentor.Assets.Controls.Windows;assembly=FrontendMentor.Assets"
                     xmlns:properties="clr-namespace:FrontendMentor.QrCodeComponent.Properties"
                     xmlns:resourceDictionaries="clr-namespace:FrontendMentor.Assets.ResourceDictionaries;assembly=FrontendMentor.Assets"
-                    xmlns:windows="clr-namespace:FrontendMentor.Assets.Controls.Windows;assembly=FrontendMentor.Assets"
-                    xmlns:windows1="clr-namespace:FrontendMentor.QrCodeComponent.Controls">
+                    xmlns:windows="clr-namespace:FrontendMentor.QrCodeComponent.Controls.Windows">
 
     <ResourceDictionary.MergedDictionaries>
         <resourceDictionaries:SharedResourceDictionary Source="pack://application:,,,/FrontendMentor.Assets;component/Styles/Controls/Generic/Windows/FrontendMentorWindowStyles.xaml" />
@@ -11,7 +11,7 @@
         <resourceDictionaries:SharedResourceDictionary Source="pack://application:,,,/FrontendMentor.QrCodeComponent;component/Fonts/Fonts.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
-    <Style BasedOn="{StaticResource {x:Type windows:FrontendMentorWindow}}" TargetType="{x:Type windows1:QrCodeComponentWindow}">
+    <Style BasedOn="{StaticResource {x:Type assetsWindows:FrontendMentorWindow}}" TargetType="{x:Type windows:QrCodeComponentWindow}">
         <Setter Property="Background" Value="{StaticResource QrCodeComponentWindow.Background.Brush}" />
         <Setter Property="FontFamily" Value="{StaticResource Outfit.FontFamily}" />
         <Setter Property="Foreground" Value="{StaticResource QrCodeComponentWindow.Foreground.Brush}" />
